+++ api/profile.ts
@@ -43,94 +43,6 @@ export const addProfileToDB = async (profile: Profile) => {
   }
 };
 
-export const addProfileWithRequestToDB = async (
-  profile: Profile,
-  request: HouseholdRequest,
-) => {
-  const profileCollectionRef = collection(db, "profiles");
-  const requestCollectionRef = collection(db, "requests");
-
-  try {
-    const docProfileRef = await addDoc(profileCollectionRef, {});
-
-    profile.id = docProfileRef.id;
-
-    console.log(
-      "Dokumentreferens id:",
-      docProfileRef.id,
-      " och profilens id:",
-      profile.id,
-    );
-
-    await updateDoc(docProfileRef, profile as Partial<Profile>);
-    const profileDoc = await getDoc(docProfileRef);
-    if (profileDoc.exists()) {
-      const profileData = profileDoc.data();
-      console.log("profilen som skapades i databasen: ", profileData);
-
-      const docRequestRef = await addDoc(requestCollectionRef, {});
-      request.id = docRequestRef.id;
-      request.profileId = profile.id;
-
-      await updateDoc(docRequestRef, request as Partial<HouseholdRequest>);
-      const requestDoc = await getDoc(docRequestRef);
-      console.log("förfrågan som skapades i databasen: ", requestDoc);
-      return requestDoc;
-    } else {
-      console.error("Uppgiftsdokumentet finns inte i databasen.");
-      return null;
-    }
-  } catch (error) {
-    console.error("Fel vid tillägg av profil:", error);
-    return null;
-  }
-};
-
-export const getRequestByHouseholdIdFromDb = async (householdId: string) => {
-  try {
-    const requestCollectionRef = collection(db, "requests");
-
-    const q = query(
-      requestCollectionRef,
-      where("householdId", "==", householdId),
-    );
-
-    const querySnapshot = await getDocs(q);
-
-    const requests: HouseholdRequest[] = [];
-
-    querySnapshot.forEach((doc) => {
-      requests.push(doc.data() as HouseholdRequest);
-    });
-
-    console.log("Förfrågningar hämtade:", requests);
-    return requests as HouseholdRequest[];
-  } catch (error) {
-    console.error("Fel vid uppdatering av hushållsidt i databasen:", error);
-    return [];
-  }
-};
-
-export const acceptProfileToHousehold = async (
-  requestId: string,
-  householdId: string,
-) => {
-  const requestDocRef = doc(db, "requests", requestId);
-
-  try {
-    //uppdaterar hushållsidt
-    await updateDoc(requestDocRef, {
-      householdId: householdId,
-    });
-
-    console.log("hushållsidt har uppdaterats i databasen.");
-    return { success: true };
-  } catch (error) {
-    console.error("Fel vid uppdatering av hushållsidt i databasen:", error);
-    return { success: false, error };
-  }
-};
-
 export const saveProfileNameToDatabase = async (
   profileId: string,
   newProfileName: string,
@@ -210,6 +122,7 @@ export const getAllProfilesByHouseholdId = async (householdId: string) => {
     const q = query(
       profileCollectionRef,
       where("householdId", "==", householdId),
+      where("status", "==", "pending"),
     );
 
     const querySnapshot = await getDocs(q);
